% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{corrvecchia_knownCovparms}
\alias{corrvecchia_knownCovparms}
\title{Vecchia approximation with Euclidean and correlation-based distances}
\usage{
corrvecchia_knownCovparms(locs, m, ordering = "maxmin",
  coordinate = NULL, dist.ordering = "correlation",
  dist.conditioning = "correlation", covmodel, covparms = NULL)
}
\arguments{
\item{locs}{A matrix with \code{n} rows and \code{p} columns. Each row of locs gives a point in \code{[0, 1]^p}}

\item{m}{Number of nearby points to condition on (the size of conditioning sets)}

\item{ordering}{"coord" or "maxmin."
If \code{ordering} is "coord," then coordinate-based ordering method is used to order the locations.
If \code{ordering} is "maxmin," then maxmin ordering method is used to order the locations.}

\item{coordinate}{a numeric vector of coordinates}

\item{dist.ordering}{"euclidean" or "correlation."
If \code{dist.ordering} is "euclidean," then euclidean distance is used to order the locations.
If \code{dist.ordering} is "correlation," then correlation based distance 1-rho is used to order the locations.}

\item{dist.conditioning}{"euclidean" or "correlation."
If \code{dist.conditioning} is "euclidean," then euclidean distance is used to construct conditioning sets.
If \code{dist.conditioning} is "correlation," then correlation based distance 1-rho is used to construct conditioning sets.}

\item{covmodel}{If \code{covmodel} is a function, then \code{covmodel} is a covariance function.
If \code{covmodel} is a matrix, then \code{covmodel} is a covariance matrix.
Please use \code{covparms = c(1)} if \code{covmodel} is a correlation matrix.}

\item{covparms}{A numerical vector with covariance parameters. It must be compatible with the argument \code{covmodel}. At \code{NULL} by default}
}
\value{
An object that specifies the Vecchia approximation for later use in likelihood evaluation or prediction. We are doing research on this.
}
\description{
The correlation-based Vecchia approximation is nothing but the Vecchia approximation with a correlation-based distance.
             It is equivalent to the Vecchia approximation with Euclidean distance for isotropic covariance function cases which are popular in application.
             If offers an automatic strategy even when Euclidean distance is not applicable (e.g. text data).
}
\examples{
n    <- 15^2
m    <- 10
locs <- matrix(runif(n * 2, 0, 1), n, 2)

covparms   <- c(1, 0.1, 10)
sigma      <- cov_expo_aniso(locs = locs, covparms = covparms)

out.euclidean <- corrvecchia_knownCovparms(locs = locs, m = m, ordering = "maxmin", coordinate = NULL, dist.ordering = "euclidean", dist.conditioning = "euclidean", covmodel = cov.aniso, covparms = covparms)
out.correlation <- corrvecchia_knownCovparms(locs = locs, m = m, ordering = "maxmin", coordinate = NULL, dist.ordering = "correlation", dist.conditioning = "correlation", covmodel = cov.aniso, covparms = covparms)

out.euclidean$ord
out.correlation$ord
}
\references{
Katzfuss, Matthias, and Joseph Guinness. "A general framework for Vecchia approximations of Gaussian processes." arXiv preprint arXiv:1708.06302 (2017).
}
