% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{cov_expo_aniso}
\alias{cov_expo_aniso}
\title{Anisotropic exponential covariance function}
\usage{
cov_expo_aniso(locs, covparms)
}
\arguments{
\item{locs}{A matrix with \code{n} rows and \code{p} columns. Each row of locs gives a point in R^d}

\item{covparms}{A vector with covariance parameters in the form (variance, range, degree of anisotropy)}
}
\value{
A matrix with \code{n} rows and \code{n} columns, with the (i, j) entry containing the anisotropic exponenital covariance between observations \code{locs[i, ]} and \code{locs[j, ]}
}
\description{
From a location matrix (\code{locs}) and a vector with covariance parameters (\code{covparms}), this function returns an anisotropic exponential covariance matrix.
}
\section{Parametrization}{
 The covariance parameter vector is (variance, range, degree of anisotropy) = \eqn{(\sigma^2 , r, \alpha)}. The form of the covariance is \deqn{ C(x, y) = \sigma^2 exp( || A ( x - y ) || / r)} where \eqn{x} and \eqn{y} are locations in R^p and A is a diagonal matrix \eqn{diag( \sqrt \alpha , 1 , ... , 1 )}.
}

\examples{
# grid locations
cov_expo_aniso(locs = expand.grid(c(0.25, 0.75), c(0.25, 0.75)), covparms = c(1, 0.1, 10))

# randomly selected locations
cov_expo_aniso(locs = matrix(runif(8), 4, 2), covparms = c(1, 0.1, 10))
}
