% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellany.R
\name{kldiv}
\alias{kldiv}
\title{Kullback Leiber (KL) divergence between two multivariate normal distributions}
\usage{
kldiv(covmat0, covmat1, mu0 = rep(0, nrow(covmat0)), mu1 = rep(0,
  nrow(covmat0)))
}
\arguments{
\item{covmat0}{A covariance matrix of the first multivariate normal distribution}

\item{covmat1}{A covariance matrix of the second multivariate normal distribution}

\item{mu0}{A mean vector of the first multivariate normal distribution}

\item{mu1}{A mean vector of the second multivariate normal distribution}
}
\value{
KL divergence between \code{N(mu0,covmat0)} and \code{N(mu1,covmat1)}
}
\description{
From two mean vectors (\code{mu0}, \code{mu1}) and two covariance matrices (\code{covmat0}, \code{covmat1}), this function returns KL divergence between \code{N(mu0,covmat0)} and \code{N(mu1,covmat1)}.
}
\examples{
covmat0 <- diag(3)
covmat1 <- matrix(c(1, 0.9, 0.7, 0.9, 1, 0.4, 0.7, 0.4, 1), 3, 3)

kldiv(covmat0 = covmat0, covmat1 = covmat1)
}
