% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_simulation.R
\name{simulate_gp_brute}
\alias{simulate_gp_brute}
\title{Naive brute force simulation of gaussian processes}
\usage{
simulate_gp_brute(n, p, loc.method = "random", covmodel,
  covparms = NULL, nugget = 0)
}
\arguments{
\item{n}{the number of locations}

\item{p}{the dimension of domain}

\item{loc.method}{the method to choose locations. The default method \code{"random"} uses randomly selected locations. The alternative \code{"grid"} uses grids as locations.}

\item{covmodel}{covariance model}

\item{covparms}{covariance parameters as a vector}

\item{nugget}{either a single (constant) nugget or a vector of nugget terms for the observations}
}
\value{
\code{simulate_gp_brute} returns
\itemize{
      \item{\code{n}: }{the number of locations}
      \item{\code{p}: }{the dimension of domain}
      \item{\code{locs}: }{a \code{n} x \code{p} matrix of locations. Each row of \code{locs} contains a location.}
      \item{\code{sigma}: }{covariance matrix with respect to \code{locs}}
      \item{\code{y}: }{a \code{n}-dimensional vector of the user-specific gaussian process generated at \code{locs}}
}
}
\description{
Naive brute force simulation of gaussian processes
}
\examples{
out <- simulate_gp_brute(n = 5^2, p = 2, loc.method = "grid", covmodel = diag(5^2), covparms = NULL, nugget = 0)
out
}
