% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_simulation.R
\name{simulate_gp_brute}
\alias{simulate_gp_brute}
\title{Brute-force simulation of gaussian processes}
\usage{
simulate_gp_brute(locs = NULL, n = NULL, p = NULL,
  meanmodel = function(loc, meanparms) 0, meanparms = NULL,
  covmodel = function(loc1, loc2, covparms) ifelse(identical(loc1, loc2),
  1, 0), covparms = NULL, seed = NULL)
}
\arguments{
\item{locs}{locations matrix or how to select locations from the unit box}

\item{n}{number of locations}

\item{p}{dimension of the unit box}

\item{meanmodel}{mean model}

\item{meanparms}{mean parameters as a vector}

\item{covmodel}{covariance model}

\item{covparms}{covariance parameters as a vector}

\item{seed}{random seed}
}
\value{
\code{simulate_gp_brute} returns a list containing
\itemize{
      \item{seed: }{random seed}
      \item{n: }{number of locations}
      \item{p: }{dimension of the unit box}
      \item{locs: }{a \code{n} x \code{p} matrix of locations}
      \item{meanvec: }{mean values at locations}
      \item{covmat: }{covariance matrix}
      \item{y: }{generated gaussian process at locations}
}
}
\description{
Brute-force simulation of gaussian processes
}
\examples{
# The following arguments give the exact same result:

simulate_gp_brute(locs = 'random', n = 3, p = 2, meanmodel = function(loc, meanparms) 0, meanparms = NULL, covmodel = function(loc1, loc2, covparms) ifelse(identical(loc1, loc2), 1, 0), covparms = NULL, seed = 1)$y

simulate_gp_brute(locs = 'random', n = 3, p = 2, meanmodel = matrix(0, 3, 1), meanparms = NULL, covmodel = diag(3), covparms = NULL, seed = 1)$y

simulate_gp_brute(locs = 'random', n = 3, p = 2, meanmodel = matrix(0, 1, 3), meanparms = NULL, covmodel = diag(3), covparms = NULL, seed = 1)$y

simulate_gp_brute(locs = 'random', n = 3, p = 2, meanmodel = as.data.frame(matrix(0, 3, 1)), meanparms = NULL, covmodel = as.data.frame(diag(3)), covparms = NULL, seed = 1)$y

simulate_gp_brute(locs = 'random', n = 3, p = 2, meanmodel = as.data.frame(matrix(0, 1, 3)), meanparms = NULL, covmodel = as.data.frame(diag(3)), covparms = NULL, seed = 1)$y

simulate_gp_brute(locs = 'random', n = 3, p = 2, meanmodel = matrix(0, 3, 1), meanparms = NULL, covmodel = rep(1, 3), covparms = NULL, seed = 1)$y

simulate_gp_brute(locs = 'random', n = 3, p = 2, meanmodel = matrix(0, 1, 3), meanparms = NULL, covmodel = 'identity', covparms = NULL, seed = 1)$y
}
